Add part drawing size limit of 1 MB
Not sure how much space is on the NerdyParts server, so
I set max drawing size to 1 MB. This should be plenty for
normal drawings, unless it's a PDF containing more than
about 10-15 separate drawings in a single file, the PDF
has embedded images, or the page size is set to something
stupid like 200 by 200.

With even 10GB free on NerdyParts, this should let us keep
about 10,000 drawings, so we shouldn't need to worry about
space any time soon.